requirements.txt
streamlit
torch
torchvision
Pillow
numpy
matplotlib
imageio
